HAPI bug workaround, Parser runs into loop while encoding Bundles
HAPI resolves References of Entries in Bundles while parsing a Bundle.
While encoding the same Bundle afterwards HAPI runs into a endless loop
if the contained resource reference themselves. E.g. a Bundle with
Organization and Endpoint Resource where the Endpoint is referenced by
the Organization and the Organization is referenced by the Endpoint.