reworked binary dao implementation, added rest integration tests
This new binary dao implementation and the underlying abstract classes
have less duplicated code and a cleaner abstraction. For CRUD operations
the PreparedStatementFactory contains all methods that need to be
specially implemented for a resource with additional content columns.
Modification of include resources is now delegated to the
search-parameter responsible for the included resource(s). Modification
of the search resource is handled by the modifySearchResultResource
method in the dao implementation.

The AbstractIntegrationTest class can be used to implement test via the
WebserviceClient against the REST interface.

The BinaryIntegrationTest includes tests for simple create, update and
search all (no search parameter).

The OrganizationIntegrationTest includes tests for search all (no search
parameter) and a search all with include endpoints.