Fixes non resolved references in websocket resources for transactions
References in resources created via a transaction- or batch-bundle are
resolved within the ResolveReferencesCommand. This command does not
change the current state for resources in memory but rather updates the
database. Leaving created and updated resources in memory in a non
resolved state. For batch and transaction result bundles this is of no
consequence, since the resources are not returned for PUT and POST
operations. For subscription websockets this was a bug.

This commit also fixes a few leftover DomainResource types and replaces
them with Resource.