New EventManager implementation for transaction-bundles
This new EventManager implementation should fix the travis build
problem.

Problem analysis: While executing transaction-bundles with POST and PUT
operations, all resources are created/updated within a DB transaction.
The old implementation, did not wait for the transaction to be
successfully committed before handling subscription events. The
EventManagerImpl has a list of currently active Subscriptions and
refreshes this list if an create/update/delete event for a Subscription
resource is handled. Refreshing the list while the transaction creating
the Subscription is not committed leads to list of current active
subscriptions missing the no committed one. If there is no active
subscription with the requested id in the EventManager, websockets can't
connect because the EventManager can't find the scubscription id in the
cache.

Thanks to @wetret for reporting the problem.