enables tasks without business key to start new process instances
The business key within a Task resource should only be used to start
process instances that need to be executed with this specific business
key, which typically means that the executed process is called from a
remote process and expected to call-back the initiating process. For
these call-backs usually a correlation key is used to handle 1:n
communication relationships.