added missing ttp endpoints to external fhir bundle of 3 medic/ttp test
entries are necessary because the first white-list update is triggered
by a literal external reference inside a task resource which needs to be
resolved against the ttp fhir endpoint. Resolving literal external
references requires a existing endpoint resource in oder for the
reference resolve to create a http client, for the necessary http HEAD
call to the remove server.