Logical references no longer replaced with literal references
Logical references are now resolved and stored as literal references in
the DB to be used with _include and _revinclude, but returned as logical
references. See ReferenceCleaner.

ReferenceExtractor now also handles input and output fields of Task
resources, as well as extensions of all resources; including extensions
of extensions. Adds test to check extensions of extensions.

Updated StructureDefinitions and ActiviyDefinitions for the simple
feasibility process.

Some code cleanup. TestSuite fixes.