ReferenceCleaner implementation to workaround a hapi parser bug
When the hapi parser, parses Bundle resources that contain resources
referencing each other, the parses adds referenced resources it can
resolve as a "resource" property in the references. If Bundles modified
in this way are serialized back into xml or json the parser runs into a
StackOverflowError.

The added ReferenceCleaner removes the added objects from the references
since they are never used anyways.