removed non mandatory properties and added profiles
Added HiGHmed organization and endpoint profiles to the test and ttp
bundles. Not added to the medic bundles, to keep them minimal. Also
removed a few properties that are not mandatory in the base profiles and
not needed. The list of HiGHmed organization and endpoint resources gets
updated via the whitelist update mechanism from the ttp with all
properties that are mandatory in the HiGHmed profiles.