Fix Docker Test FHIR Server Base
The server base was set to https://localhost/fhir but the actual hostname of the FHIR server inside the Docker network is fhir and not localhost.

The server base is used to validate the fullUrl of transaction/batch bundles on update commands. In case one restarts the BPE, it tries to update resources like CodeSystem. For that update it uses https://fhir/fhir as base URL which doesn't match https://localhost/fhir.

Steps to reproduce:

1. Follow https://github.com/highmed/highmed-dsf/wiki/Build-and-Test-Project#manual-integration-testing-local-with-docker
2. stop the BPE and restart it