fixed/implemented conditional delete and create in same bundle
Conditionally deleting an existing resource and creating a new resource
if non exists with the same condition, did not work. This fixes the
create/delete commands. Adds an integration test case to test this with
a ValueSet.
Also adds a NullPointerException protection to the preExecute of the
CreateCommand and UpdateCommand if the bundle entry has no fullUrl.