improved error handling, rearranged methods
A NullPointerExcepton could occur if a process for the given
Task.instantiatesUri could not be found. This can be the case, if a bad
ActivityDefinition exists and allows a Task to be created without the
corresponding process existing.

Since resulting NullPointerException does not have a message (return
null for .getMessage()) the updated Task (status failed) does not comply
with the base Task profile. The Task.output.value will now be always
set, since it will start with the Exception class name.

This change is inspired by #231