some documentation and code improvements, removed filePropertySupported
Remove the filePropertySupported field from the Documentation
annotation, because its value can be calculated. All properties ending
with .password are supported to be read from files using
....password.file or ..._PASSWORD_FILE

The DocumentationGenerator now moves an existing md file to md.backup
before writing the new file.

Default and example values are now written as `monospace` in the md.