allows process plugins to fail during startup without crashing the BPE
BeanCreationExceptions due to errors during initialization of spring
beans in plugins are now caught and an empty (aka no beans) application
context is used for a failed process plugin. Because of this,
JavaDelegates for BPMN Tasks may not exist. This will now result in a
warning, not a crash of the BPE.

fixes #335