new json object mapper configuration, only explicit declarations allowed
With the new json object mapper configuration auto detection of
creators, fields, getters, is-getters and setters is disabled. Elements
need to be explicitly annotated for example using @JsonCreator or
@JsonProperty.

For testing serialization of BPE variables, process plugins can use the
new org.highmed.dsf.fhir.json.ObjectMapperFactory to create a
ObjectMapper that is configured equally to the one used within the DSF
BPE.

This change means that developers of process plugins using custom BPE
variables need to add omitted json annotations in order for the
variables to sill work.

closes #339