fixed regression, unable to execute different version of same process
Since the camunda bpe uses internal version numbers to determine the
"latest version" independent of the processes versionTag, process
definition queries can not be defined using versionTag(...) and
latestVersion() in the same query. latestVersion() will always select
the latest deployed version.

The now fixed implementation only selects active processes with the
given process key and version, the resulting list is sorted by the
internal version number and the process definition with the highest
version is selected.