EndListener now added on "start" of EndEvents
In order for process plugin developers to be able to execute EndEvent
execution listeners after the process finishes and the Task resource is
written to the DSF FHIR server, the DSF EndListener is now added to the
"start" phase of EndEvents.

Execution Listeners configured in BPMN on the "start" phase will execute
before the DSF EndListener. Configured for the "end" phase, they will
execute after the DSF EndListener.